'Add the Nested routing for user forms like login, signup and forgot password, remove unnecessary css code from App.css and make the code responsive using tailwind only
