+++ Backend/config.env
@@ -3,4 +3,4 @@ RESET_SECRET=
 EMAIL_USER=a3064e124cf38e
 EMAIL_PASSWORD=56a346bed412a7
 EMAIL_HOST=sandbox.smtp.mailtrap.io
-EMAIL_PORT=5000
+EMAIL_PORT=465
